refactor(LC0091): centralize translation-ID reflection in ALCops.Common
Extract the SDK-version-compat reflection for computing XLIFF trans-unit IDs
out of the TranslatableTextShouldBeTranslated analyzer into a dedicated
ALCops.Common.Reflection.TranslationIdHelper. This keeps all SDK-version-compat
reflection in one place (alongside SymbolHelper, EnumProvider), so future SDK
bumps only require changes in Common. The analyzer now calls
TranslationIdHelper.ComputeTranslationId and retains only translation policy
(which symbols to inspect, isLabelConst decisions, null-guarding).

No behavior change: LC0091 tests unchanged (14 passed, 2 namespace tests
skipped on SDK 18.0.36; validated green on 18.0.38.52553 previously).
